requirements.txt
# Telegram scraping
telethon==1.34.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0

# HTTP client
httpx>=0.26.0

# Language detection
langdetect>=1.0.9

# Scheduling
apscheduler>=3.10.4

# Environment variables
python-dotenv>=1.0.0

# Async utilities
asyncio-throttle>=1.0.2
